Add word-source module.
Copy the word-source module from the word-source-cljs project. Change it
to fit as a library. This change allows me to call it's functions within
a node REPL.

To run it, one must execute the command `npm install sprintf-js`;
however, I do not believe I actually need this node package for this
module.