Repair issues
Previous commits had at least two issues. First, if one specified the
`top-f` argument to `typical-stage-extent`, the code failed because it
assumed that the value returned by evaluation `top-f` was a measurement
magnitude, but it was now a sequence of a magnitude and a unit.

Second, repairing the first issue uncovered a broader issue. The helper
function `as-f` returned a function that failed if the `from` and `to`
units were identical.